Extract read/write from specific modules to io
- Provide these stream related callables as functions on the io module
  that takes the relevant callables for the production of the target
  items.
